fix: update data fields to use a different data type for better handling of precision

+++ src/main/java/com/rifushigi/nomisma/entity/Country.java
@@ -1,9 +1,13 @@
 package com.rifushigi.nomisma.entity;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.data.annotation.LastModifiedDate;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Entity
@@ -31,10 +35,10 @@ public class Country {
     private String currencyCode;
 
     @Column(name = "exchange_rate")
-    private Double exchangeRate;
+    private BigDecimal exchangeRate;
 
     @Column(name = "estimated_gdp")
-    private Double estimatedGdp;
+    private BigDecimal estimatedGdp;
 
     @Column(name = "flag_url")
     private String flagUrl;
